refactor: extract onIterationEnd compaction callback from app.tsx
Move the ~100-line mid-turn compaction logic into src/app/compaction.ts
as a standalone async function. App keeps a thin useCallback wrapper.

Also exports gatewayFromConfig from app.tsx for reuse by extracted modules.
Zero behavioral change — tests and typecheck pass.

Commit-Queue-Session: cq_20260508160558_dad111b7260fef0bcd71f2b9
Coding-Agent: pi
Coding-Agent-Session: pi-019e06b6-830e-72ca-8378-e2ba2d3209c0
