+++ src/app.tsx
@@ -179,6 +179,7 @@ import {
   initMcp as initMcpFn,
   initLsp as initLspFn,
 } from "./app/mcp-lsp-init.js";
+import { onIterationEnd as onIterationEndFn } from "./app/compaction.js";
 
 type ActivePicker =
   | { kind: "file"; anchor: number; selected: number }
@@ -244,7 +245,7 @@ export interface Cfg {
   cloudToken?: string;
 }
 
-function gatewayFromConfig(cfg: Cfg): AiGatewayOptions | undefined {
+export function gatewayFromConfig(cfg: Cfg): AiGatewayOptions | undefined {
   if (!cfg.aiGatewayId) return undefined;
   return {
     id: cfg.aiGatewayId,
@@ -1162,99 +1163,19 @@ function App({
       messages: ChatMessage[],
       signal: AbortSignal,
     ): Promise<ChatMessage[]> => {
-      if (signal.aborted) return messages;
-      if (!shouldCompact({ messages })) return messages;
-
-      if (compiledContextRef.current) {
-        const store = artifactStoreRef.current;
-        const result = compactCompiled({
-          messages,
-          state: sessionStateRef.current,
-          store,
-        });
-        if (result.metrics.rawTurnsRemoved > 0) {
-          sessionStateRef.current = result.newState;
-          setEvents((e) => [
-            ...e,
-            {
-              kind: "info",
-              key: mkKey(),
-              text: `auto-compacted: ${result.metrics.estimatedTokensBefore} → ${result.metrics.estimatedTokensAfter} tokens (${result.metrics.archivedArtifacts} artifacts)`,
-            },
-          ]);
-          await saveSessionSafe();
-        }
-        // After compaction, recall memories so the model retains durable anchors
-        const manager = memoryManagerRef.current;
-        if (manager && !signal.aborted) {
-          try {
-            const cwd = process.cwd();
-            const queryText = sessionStateRef.current.task || cwd;
-            const results = await manager.recall({
-              text: queryText,
-              repoPath: cwd,
-              limit: 5,
-            });
-            if (results.length > 0 && !signal.aborted) {
-              const text = await manager.synthesizeRecalled(results);
-              const lastSystemIdx = result.newMessages.findLastIndex(
-                (m) => m.role === "system",
-              );
-              const insertIdx =
-                lastSystemIdx >= 0
-                  ? lastSystemIdx + 1
-                  : result.newMessages.length;
-              result.newMessages.splice(insertIdx, 0, {
-                role: "system",
-                content: text,
-              });
-              setEvents((e) => [
-                ...e,
-                {
-                  kind: "memory",
-                  key: mkKey(),
-                  text: `recalled ${results.length} memory${results.length === 1 ? "" : "ies"} after compaction`,
-                },
-              ]);
-              await saveSessionSafe();
-            }
-          } catch {
-            // Non-fatal
-          }
-        }
-        return result.newMessages;
-      }
-
-      // Non-compiled context: fall back to LLM summarizer
-      if (cfg && !signal.aborted) {
-        try {
-          const result = await compactMessages({
-            accountId: cfg.accountId,
-            apiToken: cfg.apiToken,
-            model: cfg.model,
-            messages,
-            signal,
-            gateway: gatewayFromConfig(cfg),
-          });
-          if (result.replacedCount > 0) {
-            setEvents((e) => [
-              ...e,
-              {
-                kind: "info",
-                key: mkKey(),
-                text: `auto-compacted: ${result.replacedCount} messages summarized`,
-              },
-            ]);
-            await saveSessionSafe();
-          }
-          return result.newMessages;
-        } catch {
-          // Non-fatal: if compaction fails, continue with original messages
-        }
-      }
-      return messages;
+      return onIterationEndFn(
+        messages,
+        signal,
+        cfg,
+        compiledContextRef,
+        artifactStoreRef,
+        sessionStateRef,
+        memoryManagerRef,
+        appendEvent,
+        saveSessionSafe,
+      );
     },
-    [cfg],
+    [cfg, saveSessionSafe, appendEvent],
   );
 
   useInput((inputChar, key) => {
